Add restart feature for Poisson-Nernst-Planck
We use binary files for phi, c_plus and c_minus.
They are written to the output folder and read from there
if asked by the user in lb.in with input tag RestartPNP.

Pay attention that by default RestartPNP is TRUE.
It means that by default we call PNP routine.